continue

+++ src/main/java/com/fem/entities/EtabInfo.java
@@ -6,7 +6,7 @@ import com.googlecode.objectify.annotation.Id;
 @Entity
 public class EtabInfo {
     @Id
-    public Integer id;
+    public long id;
 
     public String name;
     public String email;
